reparaciones de la parte de aministracion para subir imagenes

+++ admin/ajax/DelArticle.php
@@ -7,14 +7,22 @@ $iID = $_POST["iID"];
 $sDirImagen = $bd->SearchById($iID);
 $iResultado = -1;
 if ($sDirImagen[0]["TYPEMEDIA"] == 0) {
-    $bRespuestaArchivoEliminado = unlink(ARTICLE . "/" . $sDirImagen[0]["URL"]);
-    if ($bRespuestaArchivoEliminado) {
-        $iResultado = $bd->Del($iID);
+    if ((ARTICLE . "/" . $sDirImagen[0]["URL"]) != ARTICLE . "/" ) {
+        if (file_exists(ARTICLE . "/" . $sDirImagen[0]["URL"]) == true) {
+            $bRespuestaArchivoEliminado = unlink(ARTICLE . "/" . $sDirImagen[0]["URL"]);
+            if ($bRespuestaArchivoEliminado) {
+                $iResultado = $bd->Del($iID);
+            } else {
+                $bd->Desactive($iID);
+            }
+        } else {
+            $iResultado = $bd->Del($iID);
+        }
     } else {
-        $bd->Desactive($iID);
+        $iResultado = $bd->Del($iID);
     }
 } else {
-     $iResultado = $bd->Del($iID);
+    $iResultado = $bd->Del($iID);
 }
 
 $json = new stdClass();
